% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_schema_versions.R
\name{batch_download_deposit_versions}
\alias{batch_download_deposit_versions}
\title{Batch download deposit versions}
\usage{
batch_download_deposit_versions(df = list_deposit_versions(), dir_path)
}
\arguments{
\item{df}{Data frame. Has the same structure as the output of \code{\link[=list_deposit_versions]{list_deposit_versions()}}.
Default is \code{list_deposit_versions()} so that it downloads all versions of the deposit.}

\item{dir_path}{Character. Path to folder where files should be downloaded.}
}
\value{
List of download locations.
}
\description{
This is \code{download_deposit_version} wrapped in a \code{purr::pmap} call.
}
\examples{
\dontrun{
# download all versions
batch_download_deposit_versions( dir_path = "data")
}


}
