% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/becker_disease_data.R
\docType{data}
\name{becker_disease_data}
\alias{becker_disease_data}
\title{Becker et al. dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 21 columns.
}
\source{
\url{https://pharos.viralemergence.org/projects/?prj=prjRPayEvMecN}
}
\usage{
becker_disease_data
}
\description{
A bat coronavirus dataset that conforms to the wildlife disease data standard.
See data standard for field descriptions
}
\seealso{
Other Data: 
\code{\link{becker_project_metadata}},
\code{\link{disease_data_required_fields}},
\code{\link{minimal_disease_data}},
\code{\link{minimal_project_metadata}},
\code{\link{project_metadata_required_fields}},
\code{\link{schema_required_fields}}
}
\concept{Data}
\keyword{datasets}
