% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/becker_project_metadata.R
\docType{data}
\name{becker_project_metadata}
\alias{becker_project_metadata}
\title{Becker et al. project metadata}
\format{
An object of class \code{list} of length 11.
}
\source{
\url{https://www.ebi.ac.uk/pride/archive/projects/PXD031075}
}
\usage{
becker_project_metadata
}
\description{
The project metadata for a bat coronavirus dataset that conforms to the wildlife disease data standard.
See data standard for field descriptions.
}
\seealso{
Other Data: 
\code{\link{becker_disease_data}},
\code{\link{disease_data_required_fields}},
\code{\link{minimal_disease_data}},
\code{\link{minimal_project_metadata}},
\code{\link{project_metadata_required_fields}},
\code{\link{schema_required_fields}},
\code{\link{wdds_to_pharos_map}}
}
\concept{Data}
\keyword{datasets}
