% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{clean_field_names}
\alias{clean_field_names}
\title{Clean Field Names}
\usage{
clean_field_names(x)
}
\arguments{
\item{x}{Data frame or other named object}
}
\value{
object with names in \code{snakecase::to_lower_camel_case} format
}
\description{
Clean Field Names
}
\examples{

df  <- data.frame("Sample ID"= 1:10, "Name"= "Fred", "Host Identification"= "Pinus strobus")

clean_field_names(df)

}
