% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_docs.R
\name{create_object_docs}
\alias{create_object_docs}
\title{Create Docs Section for a schema object}
\usage{
create_object_docs(x, idx, required_fields, schema_dir)
}
\arguments{
\item{x}{List. Schema property or definition}

\item{idx}{Name from schema property}

\item{required_fields}{Character. Vector of required fields}

\item{schema_dir}{Character. directory where the schema is stored}
}
\value{
Character formatted markdown text
}
\description{
Create Docs Section for a schema object
}
