% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_schema.R
\name{create_object_list}
\alias{create_object_list}
\title{Create a list from a schema object}
\usage{
create_object_list(x, idx, schema_dir)
}
\arguments{
\item{x}{List. Schema property or definition}

\item{idx}{Name from schema property}

\item{schema_dir}{Character. directory where the schema is stored}
}
\value{
data frame with type and name
}
\description{
Creates a data.frame with the fields name and type
}
\seealso{
Other Schema: 
\code{\link{create_schema_list}()},
\code{\link{datacite_schema}},
\code{\link{disease_data_schema}},
\code{\link{get_ref_list}()},
\code{\link{process_array_items}()},
\code{\link{project_metadata_schema}},
\code{\link{schema_properties}},
\code{\link{schema_terms}},
\code{\link{wdds_schema}}
}
\concept{Schema}
