% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_schema.R
\name{create_schema_list}
\alias{create_schema_list}
\title{Create an expanded schema object}
\usage{
create_schema_list(schema_path = the$current_schema_path)
}
\arguments{
\item{schema_path}{Character. Path to a json-schema}
}
\value{
character vector of markdown text
}
\description{
Produces nested list for the schema. This is a
recursive set of function
}
\examples{

create_schema_list()

}
