% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_required_fields.R
\docType{data}
\name{data_required_fields}
\alias{data_required_fields}
\title{Required fields in the data object}
\format{
An object of class \code{character} of length 13.
}
\usage{
data_required_fields
}
\description{
See data standard JSON file for field descriptions.
}
\keyword{datasets}
