% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_schema.R
\docType{data}
\name{disease_data_schema}
\alias{disease_data_schema}
\title{Wildlife Disease Data Standard - data}
\format{
An object of class \code{list} of length 7.
}
\usage{
disease_data_schema
}
\description{
See data standard JSON file for field descriptions.
}
\keyword{datasets}
