% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_schema_versions.R
\name{download_deposit_version}
\alias{download_deposit_version}
\title{Download deposit version}
\usage{
download_deposit_version(zenodo_id, version, latest_version, dir_path)
}
\arguments{
\item{zenodo_id}{String. ID for a Zenodo deposit. Should correspond to the version of a deposit.}

\item{version}{String. Version number/id for the deposit (e.g. v.1.1.1).}

\item{latest_version}{Logical. Indicates that the work is designated as the latest version.}

\item{dir_path}{String. Path to directory where the files should be downloaded
e.g. "inst/extdata/wdds_archive" note no trailing slash on the path.}
}
\value{
String. Path to downloaded version.
}
\description{
Downloads and extracts some version of the deposit. This function is specific
to the structure of the wdds repo.
}
\examples{
\dontshow{if (curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# list all deposit versions
list_deposit_versions()

# download the deposit

\dontrun{
download_deposit_version("15270582","v.1.0.3",TRUE,"data")
}

\dontshow{\}) # examplesIf}
}
