% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_metadata_from_doi.R
\name{download_oa_item}
\alias{download_oa_item}
\title{Rate limited download of OA items}
\usage{
download_oa_item(entity, oa_id, dir_temp = tempdir(), sleep_time = 1)
}
\arguments{
\item{entity}{Character. What kind of openalex item is it?}

\item{oa_id}{Character. ID from openalex}

\item{dir_temp}{Character. path to directory where jons is stored.}

\item{sleep_time}{Numeric. Seconds of sleep.}
}
\value{
Character. File path to json file
}
\description{
Checks if file exists in a directory, downloads the file if its not found.
Sleeps for a given amount of time to respect rate limits on openalex servers.
}
\seealso{
Other Project Metadata: 
\code{\link{expand_tidy_dfs}()},
\code{\link{extract_metadata_from_doi}()},
\code{\link{extract_metadata_oa}()},
\code{\link{generate_metadata_csv}()},
\code{\link{generate_repeat_dfs}()},
\code{\link{make_simple_df}()}
}
\concept{Project Metadata}
