% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_metadata_from_doi.R
\name{expand_tidy_dfs}
\alias{expand_tidy_dfs}
\title{Expand tidy dataframes to project metadata template format}
\usage{
expand_tidy_dfs(tidy_df, group_prefix)
}
\arguments{
\item{tidy_df}{data frame. Each row corresponds to a complete entry.}

\item{group_prefix}{character. A repeatable metadata property in the project
metadata section of WDDS. See https://viralemergence.github.io/wddsWizard/articles/schema_overview.html#project_metadata}
}
\value{
Data frame. The data frame contains the fields Group, Variable, and Value.
}
\description{
Creates a JSON-like structure in the csv that can be processed using
established workflows in this package.
}
\examples{

# a nice tidy dataset
creators_tidy <- data.frame("Name" = paste(letters[1:10],LETTERS[1:10]),
         "Given Name" = letters[1:10],
         "Family Name" = LETTERS[1:10],
         "Name Identifier" = sample(1:100,10,FALSE),
         "Affiliation" = letters[11:20],
         "Affiliation Identifier" = 11:20,
         check.names =FALSE)

# an expanded dataset that matches the template format.
creators_tidy |>
 expand_tidy_dfs(group_prefix = "Creators")



}
\seealso{
Other Project Metadata: 
\code{\link{download_oa_item}()},
\code{\link{extract_metadata_from_doi}()},
\code{\link{extract_metadata_oa}()},
\code{\link{generate_metadata_csv}()},
\code{\link{generate_repeat_dfs}()},
\code{\link{make_simple_df}()}
}
\concept{Project Metadata}
