% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_metadata_from_doi.R
\name{extract_metadata_from_doi}
\alias{extract_metadata_from_doi}
\title{Extract Project Metadata from DOI}
\usage{
extract_metadata_from_doi(doi, file_path, write_output = TRUE)
}
\arguments{
\item{doi}{String. DOI for a published work}

\item{file_path}{String. Where should the output be written?}

\item{write_output}{Logical. Should the output be written to a file?}
}
\value{
data frame. A data frame structured in the same way as the metadata template csv.
}
\description{
Some works are explicitly connected to a publication and the metadata for that publication are fairly complete.
Instead of re-writing the metadata, it would be better to extract it and transform it.
}
\examples{

doi <-"doi.org/10.1038/s41597-025-05332-x"
extract_metadata_from_doi(doi = doi,write_output=FALSE)

}
\seealso{
Other Project Metadata: 
\code{\link{expand_tidy_dfs}()},
\code{\link{extract_metadata_oa}()},
\code{\link{generate_metadata_csv}()},
\code{\link{generate_repeat_dfs}()},
\code{\link{make_simple_df}()}
}
\concept{Project Metadata}
