% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_metadata_from_doi.R
\name{extract_metadata_oa}
\alias{extract_metadata_oa}
\title{Extract Metadata from Open Alex record}
\usage{
extract_metadata_oa(doi)
}
\arguments{
\item{doi}{Character. A digital object identifier for a published work.}
}
\value{
data frame. A data frame structured in the same way as the metadata template CSV.
}
\description{
Uses the DOI for a work to extract metadata from OpenAlex - https://openalex.org/.
The OpenAlex data model does not included some fields that are part
of the wdds project metadata  related identifiers.
}
\details{
Carefully review and edit the metadata produced.

We recommend writing the metadata to a csv, editing the csv, then
processing it as demonstrated in the project metadata tutorial.
}
\examples{

doi <- "doi.org/10.1038/s41597-025-05332-x"
extract_metadata_oa(doi = doi)

}
\seealso{
Other Project Metadata: 
\code{\link{download_oa_item}()},
\code{\link{expand_tidy_dfs}()},
\code{\link{extract_metadata_from_doi}()},
\code{\link{generate_metadata_csv}()},
\code{\link{generate_repeat_dfs}()},
\code{\link{make_simple_df}()}
}
\concept{Project Metadata}
