% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_metadata_csv.R
\name{generate_metadata_csv}
\alias{generate_metadata_csv}
\title{Generate minimal project metadata template}
\usage{
generate_metadata_csv(
  file_path,
  event_based,
  archival,
  num_creators,
  num_titles,
  identifier,
  identifier_type,
  num_subjects,
  publication_year,
  rights,
  language,
  num_descriptions,
  num_fundingReferences,
  num_related_identifiers,
  write_output = TRUE
)
}
\arguments{
\item{file_path}{String. Where should the CSV file be saved?}

\item{event_based}{Logical. Whether or not research was conducted in response to a known or suspected infectious disease outbreak, observed animal morbidity or mortality, etc.}

\item{archival}{Logical. Whether samples were from an archival source (e.g., museum collections, biobanks).}

\item{num_creators}{Integer. Number of creators for a work.}

\item{num_titles}{Integer. Number of titles for a work.}

\item{identifier}{String. A unique string that identifies a resource. Should be a DOI}

\item{identifier_type}{String. Should be DOI}

\item{num_subjects}{Integer. Number of subjects. Subject, keyword, classification code, or key phrase describing the resource}

\item{publication_year}{String. Year when work was published}

\item{rights}{String. Use one of the rights identifiers found here https://spdx.org/licenses/}

\item{language}{String. The primary language of the resource.}

\item{num_descriptions}{Integer. Number of descriptions to add to the csv. All additional information that does not fit in any of the other categories. May be used for technical information or detailed information associated with a scientific instrument}

\item{num_fundingReferences}{Integer. Number of funders to add to the csv. Name and other identifying information of a funding provider}

\item{num_related_identifiers}{Integer. Number of other works you would like to link to.}

\item{write_output}{Logical. Should the file be written?}
}
\value{
data.frame
}
\description{
This function allows you to generate a minimal metadata template for your project.
You provide certain values and it generates a csv based on those values.
Any parameter that starts with num takes an integer and creates repeat entries in the metadata csv.
All other values take a string or logical input and will prepopulate that section of the metadata csv.
}
\examples{

generate_metadata_csv(file_path = "test.csv",
event_based = TRUE,
archival = FALSE,
num_creators = 10,
num_titles = 1,
identifier = "https://doi.org/10.1080/example.doi",
identifier_type = "doi",
num_subjects = 5,
publication_year = "2025",
rights = "cc-by",
language = "en",
num_descriptions = 1,
num_fundingReferences = 4,
num_related_identifiers= 5,
write_output = FALSE) # change to TRUE to write the csv

}
