% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_metadata_csv.R
\name{generate_repeat_dfs}
\alias{generate_repeat_dfs}
\title{generate_repeat_dfs}
\usage{
generate_repeat_dfs(num_groups, group_prefix, group_variables)
}
\arguments{
\item{num_groups}{Numeric. Number of groups}

\item{group_prefix}{Character. A group name}

\item{group_variables}{Character. A comma separated scalar string of variables.}
}
\value{
data frame. Structured appropriately for the metadata csv.
}
\description{
generate_repeat_dfs
}
\examples{

related_ids_df <- generate_repeat_dfs(num_groups = 5,
group_prefix = "Related Identifiers",
group_variables = "Related Identifier,Related Identifier Type,Relation Type")

}
