% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{get_entity}
\alias{get_entity}
\title{Get entity}
\usage{
get_entity(x)
}
\arguments{
\item{x}{data frame. A "long" form data frame with the fields Group, entity_id, Value, and variable.}
}
\value{
data frame in "wide" form
}
\description{
The \code{get_entity} function creates standard entities that will be easier to transform json.
}
\details{
Pivots data from long to wide and formats column names.
}
\examples{

df <- data.frame(Group = 1, entity_id = 1, Value = 1:3, Variable = letters[1:3])

get_entity(df)


}
