% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_docs.R
\name{get_ref}
\alias{get_ref}
\title{Get schema references}
\usage{
get_ref(x, schema_dir)
}
\arguments{
\item{x}{List. Must have property "$ref"}

\item{schema_dir}{Character. Directory for the current schema.}
}
\value{
List or Character. Character is only returned if an entire schema is referenced.
}
\description{
Parses $ref calls in a schema. Can retrieve internal ('"$ref":"#/definitions/someDef") or
external references ('"$ref":"schemas/datacite/datacite.json"').
}
\details{
For external references, it can handle both pointers and references to entire schemas.
This function navigates between parent and child schemas by manipulating
variables in the  package environment \code{the}.
}
\seealso{
Other Schema Docs: 
\code{\link{create_object_docs}()},
\code{\link{create_schema_docs}()},
\code{\link{get_required_fields}()},
\code{\link{increase_docs_depth}()},
\code{\link{paste_reduce}()},
\code{\link{paste_reduce_ul}()}
}
\concept{Schema Docs}
