% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_schema.R
\name{get_ref_list}
\alias{get_ref_list}
\title{Get schema references}
\usage{
get_ref_list(x, schema_dir)
}
\arguments{
\item{x}{List. Must have property "$ref"}

\item{schema_dir}{Character. Directory for the current schema.}
}
\value{
data frame with name or type.
}
\description{
Parses $ref calls in a schema. Can retrieve internal ('"$ref":"#/definitions/someDef") or
external references ('"$ref":"schemas/datacite/datacite.json"').
}
\details{
For external references, it can handle both pointers and references to entire schemas.
This function navigates between parent and child schemas by manipulating
variables in the  package environment \code{the}.
}
