% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_docs.R
\name{increase_docs_depth}
\alias{increase_docs_depth}
\title{Increase documentation depth}
\usage{
increase_docs_depth(string)
}
\arguments{
\item{string}{Character. item to be parsed}
}
\value{
character
}
\description{
Pads the left side of any list items with an extra
4 spaces
}
