% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_schema_versions.R
\name{list_deposit_versions}
\alias{list_deposit_versions}
\title{List Versions of a deposit on Zenodo}
\usage{
list_deposit_versions(parent_id = "15020049")
}
\arguments{
\item{parent_id}{String. Identifier for a Zenodo deposit with multiple versions.}
}
\value{
Data frame. The data frame contains the Zenodo id for each version of
the deposit, as well as the version name, and logical field called latest that
indicates if this is the latest version.
}
\description{
This function list all the versions of a deposit associated with a parent id.
The parent id is used to identify a set of works that are different versions
of the same work. The parent id is provided from the Zenodo API. If you
download a JSON representation of the deposit (export to json), there
will be an attribute in that json called parent that looks like
"https://zenodo.org/api/records/15020049". The 8 digit string at the end of
the url is the parent id.
}
\examples{
\dontshow{if (curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

list_deposit_versions()



\dontshow{\}) # examplesIf}
}
