% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_template.R
\name{list_wdds_templates}
\alias{list_wdds_templates}
\title{File paths for wdds templates}
\usage{
list_wdds_templates(template_file = NULL)
}
\arguments{
\item{template_file}{character. file name for a template. Default is \code{NULL}
to return template files}
}
\value{
file paths or, if path = NULL, a list of file names
}
\description{
Displays file paths for Wildlife Disease Data Standard templates
}
\details{
If path is null, displays all files in the templates folder.
}
\seealso{
Other Templates: 
\code{\link{use_wdds_template}()}
}
\concept{Templates}
