% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_metadata_from_doi.R
\name{make_simple_df}
\alias{make_simple_df}
\title{A convenience function for making non-repeating items}
\usage{
make_simple_df(property, value)
}
\arguments{
\item{property}{string. Metadata group and variable name}

\item{value}{A value for that property.}
}
\value{
data frame. A data frame that conforms to non-repeatable structure in template.
}
\description{
A convenience function for making non-repeating items
}
\examples{
language_df <- make_simple_df(property = "language", value = "fr")

}
\seealso{
Other Project Metadata: 
\code{\link{expand_tidy_dfs}()},
\code{\link{extract_metadata_from_doi}()},
\code{\link{extract_metadata_oa}()},
\code{\link{generate_metadata_csv}()},
\code{\link{generate_repeat_dfs}()}
}
\concept{Project Metadata}
