% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal_data_examples.R
\docType{data}
\name{minimal_disease_data}
\alias{minimal_disease_data}
\title{An example of minimal disease data}
\format{
An object of class \code{data.frame} with 3 rows and 13 columns.
}
\usage{
minimal_disease_data
}
\description{
This is a minimal disease data example. It is a data frame with the minimal
items required for disease data.
}
\seealso{
Other Data: 
\code{\link{becker_disease_data}},
\code{\link{becker_project_metadata}},
\code{\link{disease_data_required_fields}},
\code{\link{minimal_project_metadata}},
\code{\link{project_metadata_required_fields}},
\code{\link{schema_required_fields}}
}
\concept{Data}
\keyword{datasets}
