% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal_project_metadata.R
\docType{data}
\name{minimal_project_metadata}
\alias{minimal_project_metadata}
\title{An example of minimal project metadata}
\format{
An object of class \code{list} of length 7.
}
\usage{
minimal_project_metadata
}
\description{
This is a minimal project metadata example. Its a list with the minimal
items required for project metadata.
}
\keyword{datasets}
