% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_docs.R
\name{paste_reduce_ul}
\alias{paste_reduce_ul}
\title{Paste Reduce unordered list item}
\usage{
paste_reduce_ul(x, y, sep = "\\n - ")
}
\arguments{
\item{x}{Character}

\item{y}{Character}

\item{sep}{Character. Default is a line break followed by a dash \code{"\\n - "} to
create an unordered list in markdown.}
}
\value{
Character
}
\description{
A paste function that can be used with \code{purrr::reduce} to build up nested
documentation items
}
\examples{

text_a <- "hello"
text_b <- "world"
paste_reduce_ul(text_a,text_b)

}
\seealso{
Other Schema Docs: 
\code{\link{create_object_docs}()},
\code{\link{create_schema_docs}()},
\code{\link{get_ref}()},
\code{\link{get_required_fields}()},
\code{\link{increase_docs_depth}()},
\code{\link{paste_reduce}()}
}
\concept{Schema Docs}
