% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{prep_array}
\alias{prep_array}
\title{Prep array}
\usage{
prep_array(x)
}
\arguments{
\item{x}{a list object.}
}
\value{
unnamed vector
}
\description{
Prep array
}
\examples{

# this form can arise because of the csv template
nested_list <- list(list("formats" = list("formats" = "csv",
"formats" = "fasta")))

prep_array(nested_list)

}
