% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{prep_creators}
\alias{prep_creators}
\title{Prepare creators}
\usage{
prep_creators(x)
}
\arguments{
\item{x}{data frame or named list.}
}
\value{
List of unboxed data frames
}
\description{
The creator object can be complex so we prepare components of the final object (e.g. affiliation, nameIdentifiers) then run prep_array_objects
}
\examples{

wddsWizard::becker_project_metadata$creators |>
 prep_creators()

}
