% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{prep_data}
\alias{prep_data}
\title{Prepare Data}
\usage{
prep_data(x)
}
\arguments{
\item{x}{named vector, list, or data frame}
}
\value{
List of formatted objects
}
\description{
Prepares an object of arrays.
}
\details{
Note that unboxing will only work on items where you have 1:1 key value pair. So if you have a dataframe with multiple rows or a list  with
multiple values at a given position, it won't work.
}
\examples{

cars_small  <- datasets::cars[1:10,]

# creates an array of objects where each
# row is an object
cars_small |>
  jsonlite::toJSON(pretty = TRUE)

# creates an object with 2 arrays
prep_object(cars_small) |>
  jsonlite::toJSON(pretty = TRUE)

# this makes no difference
x <- list("hello" = 1:10, "world" = "Earth")

prep_object(x) |>
  jsonlite::toJSON(pretty = TRUE)


}
