% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{prep_descriptions}
\alias{prep_descriptions}
\title{Prepare descriptions}
\usage{
prep_descriptions(x)
}
\arguments{
\item{x}{Data frame/Tibble containing description items}
}
\value{
List with x marked as unbox (do not make an array)
}
\description{
Wrapper for \code{prep_array_objects}.
}
\examples{

x  <- wddsWizard::becker_project_metadata$descriptions

prep_descriptions(x) |> jsonlite::toJSON()


}
