% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{prep_for_json}
\alias{prep_for_json}
\title{Prepare data for json}
\usage{
prep_for_json(x, prep_methods_list = prep_methods())
}
\arguments{
\item{x}{list. Named list of data frames, lists, or vectors. For methods to be applied,
the names of the list items should match the names in the methods list}

\item{prep_methods_list}{list. Named list of methods where each items is a
function to applied to corresponding items in x. Default is full list of methods
from \code{\link[=prep_methods]{prep_methods()}}.}
}
\value{
Named list where methods have been applied.
}
\description{
Uses\code{purrr::modify_at} to apply a set of methods at specific locations in a
list.
}
\examples{

wddsWizard::becker_project_metadata |>
   prep_for_json()

 a <- list("hello_world" = 1:10 )
 methods_list <- list("hello_world" = function(x){x*2},
                       "unused_method" = function(x){x/2})
 prep_for_json(a,methods_list)


}
