% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{prep_fundingReferences}
\alias{prep_fundingReferences}
\title{Prepare funding references}
\usage{
prep_fundingReferences(x)
}
\arguments{
\item{x}{list of tibbles/data frames or a tibble/data frame}
}
\value{
list of single row unboxed data frames
}
\description{
creates an array of objects
}
\examples{

wddsWizard::becker_project_metadata$fundingReferences |>
 prep_fundingReferences()

}
