% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{prep_identifier}
\alias{prep_identifier}
\title{Prep identifier}
\usage{
prep_identifier(x)
}
\arguments{
\item{x}{data frame with identifier properties}
}
\value{
List with x marked as do not unbox
}
\description{
Prepare identifier for a scholarly work. Wrapper for prep_array_objects
}
\examples{

wddsWizard::becker_project_metadata$identifiers |> prep_identifier()

}
