% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{prep_identifiers}
\alias{prep_identifiers}
\title{Prep identifiers}
\usage{
prep_identifiers(x)
}
\arguments{
\item{x}{data frame with identifier properties}
}
\value{
List with x marked as do not unbox
}
\description{
Prepare identifiers for a scholarly work. Wrapper for prep_array_objects
}
\examples{

wddsWizard::becker_project_metadata$identifiers |> prep_identifiers()

}
