% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{prep_language}
\alias{prep_language}
\title{Prep language}
\usage{
prep_language(x)
}
\arguments{
\item{x}{named list, vector, or data.frame of with 1:1 name:value pairs}
}
\value{
an unboxed dataframe with 1 row
}
\description{
Prepare the language property - this should describe the language
of the scholarly work.
}
\examples{

a <- data.frame("language" = "en")

prep_language(a)

}
