% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{prep_methodology}
\alias{prep_methodology}
\title{Prep methodology for conversion to json}
\usage{
prep_methodology(x)
}
\arguments{
\item{x}{List. methodology component of a list}
}
\value{
properly formatted list
}
\description{
Prep methodology for conversion to json
}
\examples{

\dontrun{
prepped_list <- project_metadata_list_entities
 prepped_list$methodology <- prep_methodology(project_metadata_list_entities$methodology)

 OR

 prepped_list <- purrr::modify_at(project_metadata_list_entities,"methodology",prep_methodology)
}

}
