% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{prep_methods}
\alias{prep_methods}
\title{Prepare methods}
\usage{
prep_methods()
}
\value{
list of methods
}
\description{
Collection of methods for preparing data conveniently named to make
preparing easier
}
\examples{

prep_methods()

}
