% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{prep_publicationYear}
\alias{prep_publicationYear}
\title{Prepare publication year items}
\usage{
prep_publicationYear(x)
}
\arguments{
\item{x}{Named vector, data frame, or list}
}
\value{
an unboxed dataframe with 1 row
}
\description{
wrapper for prep atomic
}
\examples{
pub_year <- data.frame("publicationYear" = "2025")

prep_language(pub_year)


}
