% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{prep_relatedIdentifiers}
\alias{prep_relatedIdentifiers}
\title{Prepare related identifiers}
\usage{
prep_relatedIdentifiers(x)
}
\arguments{
\item{x}{data frame with related identifier properties}
}
\value{
List with x marked as do not unbox
}
\description{
Prepare related identifiers
}
\examples{

wddsWizard::becker_project_metadata$relatedIdentifiers |> prep_relatedIdentifiers()

}
\seealso{
Other JSON Prep: 
\code{\link{clean_field_names}()},
\code{\link{get_entity}()},
\code{\link{prep_affiliation}()},
\code{\link{prep_atomic}()},
\code{\link{prep_creators}()},
\code{\link{prep_descriptions}()},
\code{\link{prep_for_json}()},
\code{\link{prep_from_metadata_template}()},
\code{\link{prep_fundingReferences}()},
\code{\link{prep_identifier}()},
\code{\link{prep_language}()},
\code{\link{prep_methodology}()},
\code{\link{prep_methods}()},
\code{\link{prep_nameIdentifiers}()},
\code{\link{prep_object}()},
\code{\link{prep_publicationYear}()},
\code{\link{prep_rights}()},
\code{\link{prep_subjects}()}
}
\concept{JSON Prep}
