% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{prep_rights}
\alias{prep_rights}
\title{Prepare rights}
\usage{
prep_rights(x)
}
\arguments{
\item{x}{named list, vector, or data.frame of with 1:1 name:value pairs}
}
\value{
list of unboxed data frames
}
\description{
Prepares an array of objects
}
\examples{

 wddsWizard::becker_project_metadata$rights |> prep_rights()

}
