% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{prep_subjects}
\alias{prep_subjects}
\title{Prepare subjects}
\usage{
prep_subjects(x)
}
\arguments{
\item{x}{named list, vector, or data.frame of with 1:1 name:value pairs}
}
\value{
list of unboxed data frames
}
\description{
Subjects or keywords describing a work. Prepares an array of objects
}
\examples{
wddsWizard::becker_project_metadata$subjects |> prep_subjects()

}
