% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_metadata_required_fields.R
\docType{data}
\name{project_metadata_required_fields}
\alias{project_metadata_required_fields}
\title{Required fields in the project metadata object}
\format{
An object of class \code{character} of length 7.
}
\usage{
project_metadata_required_fields
}
\description{
See data standard JSON file for field descriptions.
}
\keyword{datasets}
