% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_schema.R
\docType{data}
\name{schema_properties}
\alias{schema_properties}
\title{Wildlife Disease Data Standard - schema properties}
\format{
An object of class \code{data.frame} with 78 rows and 4 columns.
}
\usage{
schema_properties
}
\description{
A data frame of schema names and types.
}
\seealso{
Other Schema: 
\code{\link{datacite_schema}},
\code{\link{disease_data_schema}},
\code{\link{project_metadata_schema}},
\code{\link{schema_terms}},
\code{\link{wdds_schema}}
}
\concept{Schema}
\keyword{datasets}
