% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_required_fields.R
\docType{data}
\name{schema_required_fields}
\alias{schema_required_fields}
\title{Wildlife Disease Data Standard required fields}
\format{
An object of class \code{character} of length 8.
}
\usage{
schema_required_fields
}
\description{
See data standard JSON file for field descriptions. This is a vector of
the required fields for the entire schema.
}
\keyword{datasets}
