% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_schema_versions.R
\name{set_wdds_version}
\alias{set_wdds_version}
\title{Set the wdds version for the package}
\usage{
set_wdds_version(version = "latest")
}
\arguments{
\item{version}{Character. identifier for a version e.g. "v.1.0.2" or "latest".
Default is "latest".}
}
\value{
Character. Current schema version.
}
\description{
Used to keep the package and data standard in alignment.
}
