% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_to_dcmi.R
\docType{data}
\name{spdx_licenses}
\alias{spdx_licenses}
\title{Data frame of SPDX licenses}
\format{
An object of class \code{data.frame} with 706 rows and 9 columns.
}
\source{
\url{https://github.com/spdx/license-list-data/blob/main/json/licenses.json}
}
\usage{
spdx_licenses
}
\description{
A table with SPDX license metadata. Use spdx_licenses$licenseId when uploading
data to Zenodo.
}
\seealso{
Other Data: 
\code{\link{becker_disease_data}},
\code{\link{becker_project_metadata}},
\code{\link{disease_data_required_fields}},
\code{\link{minimal_disease_data}},
\code{\link{minimal_project_metadata}},
\code{\link{project_metadata_required_fields}},
\code{\link{schema_required_fields}},
\code{\link{wdds_to_dcmi_map}},
\code{\link{wdds_to_pharos_map}}
}
\concept{Data}
\keyword{datasets}
