% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_to_dcmi.R
\name{translate_to_dcmi}
\alias{translate_to_dcmi}
\title{translate to dcmi}
\usage{
translate_to_dcmi(item, translation_map)
}
\arguments{
\item{item}{List. Item to be translated.}

\item{translation_map}{List. Instructions for translating the item}
}
\value{
List. Item that has been translated to DCMI
}
\description{
translate to dcmi
}
