% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_template.R
\name{use_wdds_template}
\alias{use_wdds_template}
\title{Use a wildlife disease data standard template}
\usage{
use_wdds_template(
  template_file = NULL,
  folder = fs::path_wd(),
  file_name = NULL,
  open = rlang::is_interactive(),
  overwrite = FALSE
)
}
\arguments{
\item{template_file}{character. File name for a template. Defaults to \code{NULL} to return all template files.}

\item{folder}{character. Where should the template be copied to? Default is the current working directory.}

\item{file_name}{character. What should the copied file be called? Default is to use whatever value is supplied to template_file.}

\item{open}{logical. Should the file be opened? Defaults to TRUE if interactive.}

\item{overwrite}{logical. Should a file with the same name in the destination folder be overwritten? Default is FALSE to avoid accidentally overwriting data.}
}
\value{
Character. If no template_file value is provided, lists all template files in the package. If a file is created, it returns the file path for that new file.
}
\description{
This function allows you to easily copy and open a template from the package.
}
\examples{

# return available templates
use_wdds_template()

\dontrun{

# makes a copy of the disease data template in the current working directory
use_wdds_template("disease_data_template.csv")
}

}
\seealso{
Other Templates: 
\code{\link{list_wdds_templates}()}
}
\concept{Templates}
