% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_schema_versions.R
\name{wdds_data_templates}
\alias{wdds_data_templates}
\title{Provides Access to Versioned Data Template Files}
\usage{
wdds_data_templates(version = NULL, file = NULL)
}
\arguments{
\item{version}{Character. Version of the wdds deposit. Leave as NULL to see
all versions.}

\item{file}{Character. Specific file from the wdds deposit. Leave as NULL to
see all files in a version.}
}
\value{
Character. Either version identifiers or file paths.
}
\description{
Since schema versions may change during the life cycle of project, it is
important that users have access to all schema versions via this package.
This function allows you to quickly retrieve whichever version of the data
templates you may need.
}
\details{
This function does three things.
\enumerate{
\item Shows all versions of the schema in the package if version is \code{NULL}
\item Provides paths to all example data files associated with a version of the schema if version is not \code{NULL} and file is \code{NULL}
\item Provides a specific file path in a specific version of the example data if both version and file are specified.
}
}
\examples{

# see which versions are in the package

wdds_data_templates()

# see files associated with a version

wdds_data_templates(version = "latest")

# get the file path for a specific file

wdds_data_templates(version = "v_1_0_2",file = "disease_data_template.csv")


}
