% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_schema_versions.R
\name{wdds_example_data}
\alias{wdds_example_data}
\title{Provides Access to Versioned Example Data Files}
\usage{
wdds_example_data(version = NULL, file = NULL)
}
\arguments{
\item{version}{Character or NULL. Version of the wdds deposit. Leave as NULL to see
all versions. Default is NULL to return a character vector of versions.}

\item{file}{Character or NULL. Specific file from the wdds deposit. Leave as NULL to
see all files in a version. Default is NULL to return all files associated with a
given version.}
}
\value{
Character. Either version identifiers or file paths.
}
\description{
Since schema versions may change during the life cycle of project, it is
important that users have access to all schema versions via this package.
This function allows you to quickly retrieve whichever version of the
example data you may need.
}
\details{
This function does three things.
\enumerate{
\item Shows all versions of the schema in the package if version is NULL.
\item Provides paths to all example data files associated with a version of the schema if version is provided and file is NULL.
\item Provides a specific file path in a specific version of the example data if both file and version are provided.
}
}
\examples{

# see which versions are in the package

wdds_example_data()

# see files associated with a version

wdds_example_data(version = "latest")

# get the file path for a specific file

wdds_example_data(version = "v_1_0_2", file = "Becker_demo_dataset.xlsx")

}
\seealso{
Other WDDS deposit: 
\code{\link{batch_download_deposit_versions}()},
\code{\link{download_deposit_version}()},
\code{\link{list_deposit_versions}()},
\code{\link{sanitize_version}()},
\code{\link{set_wdds_version}()},
\code{\link{wdds_data_templates}()},
\code{\link{wdds_json}()}
}
\concept{WDDS deposit}
