% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_schema_versions.R
\name{wdds_json}
\alias{wdds_json}
\title{Provides Access to Versioned Schema Files}
\usage{
wdds_json(version = NULL, file = NULL)
}
\arguments{
\item{version}{Character. Version of the wdds deposit. Leave as NULL to see
all versions.}

\item{file}{Character. Specific file from the wdds deposit. Leave as NULL to
see all files in a version.}
}
\value{
Character. Either version identifiers or file paths.
}
\description{
Since schema versions may change during the life cycle of project, it is
important that users have access to all schema versions via this package.
This function allows you to quickly retrieve whichever schema version you
may need.
}
\details{
This function does three things.
\enumerate{
\item Shows all versions of the schema in the package,
\item Provides paths to all schema files associated with a version of the schema
\item Provides a specific file path in a specific version of the schema.
}
}
\examples{

# see which versions are in the package

wdds_json()

# see files associated with a version

wdds_json(version = "latest")

# get the file path for a specific file

wdds_json(version = "v_1_0_2",file = "schemas/disease_data.json")


}
