% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_template.R
\name{wdds_template}
\alias{wdds_template}
\title{File paths for wdds templates}
\usage{
wdds_template(template_file = NULL)
}
\arguments{
\item{template_file}{character. file name for a template. Default is \code{NULL}
to return template files}
}
\value{
file paths or, if path = NULL, a list of file names
}
\description{
Displays file paths for Wildlife Disease Data Standard templates
}
\details{
If path is null, displays all files in the templates folder.
}
