% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_to_dcmi.R
\name{wdds_to_dcmi}
\alias{wdds_to_dcmi}
\title{WDDS to the Dublin Core Metadata Initiative}
\usage{
wdds_to_dcmi(
  metadata_to_translate,
  translation_map = wddsWizard::wdds_to_dcmi_map
)
}
\arguments{
\item{metadata_to_translate}{List. Metadata that conforms to the WDDS data standard but is not prepped for JSON. See \code{prep_from_metadata_template}}

\item{translation_map}{List. A list that describes how to translate from WDDS to DCMI.}
}
\value{
List. Translated metadata with appropriate names.
}
\description{
Converts WDDS project metadata to Zenodo flavored DCMI metadata.
}
\examples{

project_metadata <- wdds_example_data(version = "latest",
                                    file = "example_project_metadata.csv")|>
    read.csv()

test_pmd <- project_metadata |>
  prep_from_metadata_template(json_prep = FALSE)

test_pmd$rights$rights <- "CC0-1.0"

dcmi_metadata <- wdds_to_dcmi(metadata_to_translate = test_pmd,
                              translation_map =  wddsWizard::wdds_to_dcmi_map)


}
\seealso{
Other Standards Mapping: 
\code{\link{na_to_blank}()},
\code{\link{translate_to_dcmi}()},
\code{\link{wdds_to_pharos}()}
}
\concept{Standards Mapping}
