% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_to_dcmi.R
\docType{data}
\name{wdds_to_dcmi_map}
\alias{wdds_to_dcmi_map}
\title{WDDS to DCMI metadata mapping}
\format{
An object of class \code{list} of length 2.
}
\source{
\url{https://github.com/ropenscilabs/deposits/blob/main/inst/extdata/dc/schema.json}
}
\usage{
wdds_to_dcmi_map
}
\description{
A list that maps variables between WDDS and the DCMI data standards.
}
\seealso{
Other Data: 
\code{\link{becker_disease_data}},
\code{\link{becker_project_metadata}},
\code{\link{disease_data_required_fields}},
\code{\link{minimal_disease_data}},
\code{\link{minimal_project_metadata}},
\code{\link{project_metadata_required_fields}},
\code{\link{schema_required_fields}},
\code{\link{spdx_licenses}},
\code{\link{wdds_to_pharos_map}}
}
\concept{Data}
\keyword{datasets}
