% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_to_pharos.R
\name{wdds_to_pharos}
\alias{wdds_to_pharos}
\title{Convert WDDS disease data to PHAROS data}
\usage{
wdds_to_pharos(wdds_disease_data)
}
\arguments{
\item{wdds_disease_data}{Data frame. A Disease Data set that conforms to
the wdds data standard.}
}
\value{
Data frame. A tabular data set that conforms to the PHAROS data model.
}
\description{
As of 11 September 2025, WDDS and the PHAROS data model are not fully aligned.
This function converts data that conforms to WDDS into the PHAROS data model.
See \code{wdds_to_pharos_map} for the data model crosswalk.
}
\examples{

wdds_to_pharos(wdds_disease_data = wddsWizard::minimal_disease_data)

# data must be written to CSV then uploaded to PHAROS

}
\seealso{
Other Standards Mapping: 
\code{\link{wdds_to_dcmi}()}
}
\concept{Standards Mapping}
