% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_to_pharos.R
\docType{data}
\name{wdds_to_pharos_map}
\alias{wdds_to_pharos_map}
\title{WDDS to PHAROS metadata mapping}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 45 rows and 2 columns.
}
\source{
\url{https://pharos.viralemergence.org/}
}
\usage{
wdds_to_pharos_map
}
\description{
A table that maps variables between WDDS and the PHAROS data standard (11 September 2025).
Will be deprecated once PHARSO and wdds are aligned.
}
\seealso{
Other Data: 
\code{\link{becker_disease_data}},
\code{\link{becker_project_metadata}},
\code{\link{disease_data_required_fields}},
\code{\link{minimal_disease_data}},
\code{\link{minimal_project_metadata}},
\code{\link{project_metadata_required_fields}},
\code{\link{schema_required_fields}},
\code{\link{spdx_licenses}},
\code{\link{wdds_to_dcmi_map}}
}
\concept{Data}
\keyword{datasets}
