% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wddsWizard-package.R
\docType{package}
\name{wddsWizard-package}
\alias{wddsWizard}
\alias{wddsWizard-package}
\title{wddsWizard: Data Wizard for a Minimal Wildlife Disease Data Standard}
\description{
Facilitates compliance with and use of the Wildlife Disease Data Standard stored on Zenodo \url{https://doi.org/10.5281/zenodo.15020049}. It allows users to restructure and validate datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://viralemergence.github.io/wddsWizard/}
  \item Report bugs at \url{https://github.com/viralemergence/wddsWizard/issues}
}

}
\author{
\strong{Maintainer}: Collin Schwantes \email{collin@viralemergence.org} (\href{https://orcid.org/0000-0002-9882-941X}{ORCID})

}
\keyword{internal}
