% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/becker_project_metadata.R
\docType{data}
\name{becker_project_metadata}
\alias{becker_project_metadata}
\title{Becker et al. project metadata}
\format{
An object of class \code{list} of length 10.
}
\source{
\url{https://www.ebi.ac.uk/pride/archive/projects/PXD031075}
}
\usage{
becker_project_metadata
}
\description{
The project metadata for a bat coronavirus dataset that conforms to the wildlife disease data standard.
See data standard for field descriptions.
}
\keyword{datasets}
