% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{clean_field_names}
\alias{clean_field_names}
\title{Clean Field Names}
\usage{
clean_field_names(x)
}
\arguments{
\item{x}{Data frame or other named object}
}
\value{
object with names in \code{snakecase::to_lower_camel_case} format
}
\description{
Clean Field Names
}
\examples{

df <- data.frame("Sample ID" = 1:10, "Name" = "Fred", "Host Identification" = "Pinus strobus")

clean_field_names(df)

}
\seealso{
Other JSON Prep: 
\code{\link{get_entity}()},
\code{\link{prep_affiliation}()},
\code{\link{prep_array}()},
\code{\link{prep_array_objects}()},
\code{\link{prep_atomic}()},
\code{\link{prep_creators}()},
\code{\link{prep_data}()},
\code{\link{prep_descriptions}()},
\code{\link{prep_for_json}()},
\code{\link{prep_from_metadata_template}()},
\code{\link{prep_fundingReferences}()},
\code{\link{prep_identifier}()},
\code{\link{prep_language}()},
\code{\link{prep_methodology}()},
\code{\link{prep_methods}()},
\code{\link{prep_nameIdentifiers}()},
\code{\link{prep_object}()},
\code{\link{prep_publicationYear}()},
\code{\link{prep_relatedIdentifiers}()},
\code{\link{prep_rights}()},
\code{\link{prep_subjects}()},
\code{\link{prep_titles}()}
}
\concept{JSON Prep}
