% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_docs.R
\name{create_schema_docs}
\alias{create_schema_docs}
\title{Create Documentation for a schema}
\usage{
create_schema_docs(schema_path = the$current_schema_path, sep = "\\n")
}
\arguments{
\item{schema_path}{Character. Path to a json-schema}
}
\value{
character vector of markdown text
}
\description{
Produces nested markdown that documents a schema. This is a
recursive set of function
}
\examples{

create_schema_docs()

}
