% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_schema.R
\name{create_schema_list}
\alias{create_schema_list}
\title{Create an expanded schema object}
\usage{
create_schema_list(schema_path = the$current_schema_path)
}
\arguments{
\item{schema_path}{Character. Path to a json-schema. Default is the current
schema path from the package environment,}
}
\value{
character vector of markdown text
}
\description{
Produces a list of data frame with name and type for the schema.
This is a recursive set of function and may be expanded to get other properties.
}
\examples{

create_schema_list()

}
\seealso{
Other Schema: 
\code{\link{create_object_list}()},
\code{\link{datacite_schema}},
\code{\link{disease_data_schema}},
\code{\link{get_ref_list}()},
\code{\link{process_array_items}()},
\code{\link{project_metadata_schema}},
\code{\link{schema_properties}},
\code{\link{schema_terms}},
\code{\link{wdds_schema}}
}
\concept{Schema}
