% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_schema.R
\docType{data}
\name{datacite_schema}
\alias{datacite_schema}
\title{Datacite Data Stadnard}
\format{
An object of class \code{list} of length 9.
}
\usage{
datacite_schema
}
\description{
See data standard JSON file for field descriptions.
}
\seealso{
Other Schema: 
\code{\link{disease_data_schema}},
\code{\link{project_metadata_schema}},
\code{\link{schema_obj}},
\code{\link{schema_properties}},
\code{\link{schema_terms}},
\code{\link{wdds_schema}}
}
\concept{Schema}
\keyword{datasets}
