% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_data_required_fields.R
\docType{data}
\name{disease_data_required_fields}
\alias{disease_data_required_fields}
\title{Required fields in the disease data object}
\format{
An object of class \code{character} of length 9.
}
\usage{
disease_data_required_fields
}
\description{
See data standard JSON file for field descriptions.
}
\seealso{
Other Data: 
\code{\link{becker_disease_data}},
\code{\link{becker_project_metadata}},
\code{\link{minimal_disease_data}},
\code{\link{minimal_project_metadata}},
\code{\link{project_metadata_required_fields}},
\code{\link{schema_required_fields}},
\code{\link{wdds_to_pharos_map}}
}
\concept{Data}
\keyword{datasets}
