% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_metadata_from_doi.R
\name{extract_metadata_from_doi}
\alias{extract_metadata_from_doi}
\title{Extract Project Metadata from DOI}
\usage{
extract_metadata_from_doi(doi, file_path, write_output = TRUE)
}
\arguments{
\item{doi}{String. DOI for a published work}

\item{file_path}{String.}

\item{write_output}{}
}
\description{
Some works are explicitly connected to a publication and the metadata for that publication are fairly complete.
Instead of re-writing the metadata, it would be better to extract it and transform it.
}
