% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{get_entity}
\alias{get_entity}
\title{Get entity}
\usage{
get_entity(x)
}
\arguments{
\item{x}{data frame. A "long" form data frame with the fields Group, entity_id, Value, and variable.}
}
\value{
data frame in "wide" form
}
\description{
The \code{get_entity} function creates standard entities that will be easier to transform json.
}
\details{
Pivots data from long to wide and formats column names.
}
\examples{

df <- data.frame(Group = 1, entity_id = 1, Value = 1:3, Variable = letters[1:3])

get_entity(df)


}
\seealso{
Other JSON Prep: 
\code{\link{clean_field_names}()},
\code{\link{prep_affiliation}()},
\code{\link{prep_array_objects}()},
\code{\link{prep_atomic}()},
\code{\link{prep_creators}()},
\code{\link{prep_data}()},
\code{\link{prep_descriptions}()},
\code{\link{prep_for_json}()},
\code{\link{prep_from_metadata_template}()},
\code{\link{prep_fundingReferences}()},
\code{\link{prep_identifier}()},
\code{\link{prep_language}()},
\code{\link{prep_methodology}()},
\code{\link{prep_methods}()},
\code{\link{prep_nameIdentifiers}()},
\code{\link{prep_object}()},
\code{\link{prep_publicationYear}()},
\code{\link{prep_relatedIdentifiers}()},
\code{\link{prep_rights}()},
\code{\link{prep_subjects}()},
\code{\link{prep_titles}()}
}
\concept{JSON Prep}
