% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_docs.R
\name{get_required_fields}
\alias{get_required_fields}
\title{Get the required fields for a schema}
\usage{
get_required_fields(schema_list)
}
\arguments{
\item{schema_list}{List from jsonlite::read_json}
}
\value{
character vector of required fields
}
\description{
Gets the required fields for an object
}
