% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_docs.R
\name{increase_docs_depth}
\alias{increase_docs_depth}
\title{Increase documentation depth}
\usage{
increase_docs_depth(string)
}
\arguments{
\item{string}{Character. item to be parsed}
}
\value{
character
}
\description{
Pads the left side of any list items with an extra
4 spaces
}
\seealso{
Other Schema Docs: 
\code{\link{create_object_docs}()},
\code{\link{create_schema_docs}()},
\code{\link{get_ref}()},
\code{\link{get_required_fields}()},
\code{\link{paste_reduce}()},
\code{\link{paste_reduce_ul}()}
}
\concept{Schema Docs}
