% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal_data_examples.R
\docType{data}
\name{minimal_project_metadata}
\alias{minimal_project_metadata}
\title{An example of minimal project metadata}
\format{
An object of class \code{list} of length 7.
}
\usage{
minimal_project_metadata
}
\description{
This is a minimal project metadata example. It is a list with the minimal
items required for project metadata.
}
\seealso{
Other Data: 
\code{\link{becker_disease_data}},
\code{\link{becker_project_metadata}},
\code{\link{disease_data_required_fields}},
\code{\link{minimal_disease_data}},
\code{\link{project_metadata_required_fields}},
\code{\link{schema_required_fields}}
}
\concept{Data}
\keyword{datasets}
