% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_docs.R
\name{paste_reduce}
\alias{paste_reduce}
\title{Paste Reduce}
\usage{
paste_reduce(x, y, sep = "\\n")
}
\arguments{
\item{x}{Character}

\item{y}{Character}

\item{sep}{Character. Default is a line break \code{"\\n"}}
}
\value{
Character
}
\description{
A paste function that can be used with \code{purrr::reduce} to build up nested
documentation items
}
\examples{

text_a <- "hello"
text_b <- "world"
paste_reduce(text_a,text_b)

}
\seealso{
Other Schema Docs: 
\code{\link{create_object_docs}()},
\code{\link{create_schema_docs}()},
\code{\link{get_ref}()},
\code{\link{get_required_fields}()},
\code{\link{increase_docs_depth}()},
\code{\link{paste_reduce_ul}()}
}
\concept{Schema Docs}
