% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{prep_affiliation}
\alias{prep_affiliation}
\title{prep affiliation}
\usage{
prep_affiliation(x)
}
\arguments{
\item{x}{Data frame from prep_creators}
}
\value{
Data frame with affiliation fields in a list column called \code{affiliation}
}
\description{
There are affiliations associated with a creator.
}
\details{
Affiliation in datacite is an array of objects
with properties name, affiliationIdentifier, affiliationIdentifierScheme,
and schemeURI. This function takes the affiliation fields and restructures
as a list within the dataframe.

Affiliation fields to be converted to a list: "affiliation", #' "affiliationIdentifier",
"affiliationIdentifierScheme" ,
"schemeUri"
}
\examples{
creator_df <- wddsWizard::becker_project_metadata$creators[[1]]
creator_df_aff_prepped  <- prep_affiliation(creator_df)

}
