% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{prep_affiliation}
\alias{prep_affiliation}
\title{prep affiliation}
\usage{
prep_affiliation(x)
}
\arguments{
\item{x}{tibble from prep_creators}
}
\value{
tibble with affilition fields in a list column called \code{affilition}
}
\description{
There are affiliations associated with a creator.
}
\details{
Affiliation in datacite is an array of objects
with properties name, affiliationIdentifier, affiliationIdentifierScheme,
and schemeURI. This function takes the affiliation fields and restructures
as a list within the dataframe.

Affiliation fields to be converted to a list: "affiliation", #' "affiliationIdentifier",
"affiliationIdentifierScheme" ,
"schemeUri"
}
