% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{prep_atomic}
\alias{prep_atomic}
\title{Prepare atomic}
\usage{
prep_atomic(x)
}
\arguments{
\item{x}{vector}
}
\value{
an unboxed dataframe with 1 row
}
\description{
This is a thin wrapper for \code{jsonlite::unbox}. It stops \code{jsonlite} from
representing single character, numeric, logical, etc. items as arrays.
}
\details{
This is useful when a property or definition is of type string, number, logical and of length 1.
}
\examples{

x <- 1

# values in x are stored in an array
x|>
jsonlite::toJSON()
 # output is [1]

# values in x are NOT stored in an array (no square brackets)
prep_atomic(x) |>
  jsonlite::toJSON()
# output is 1

}
\seealso{
Other JSON Prep: 
\code{\link{clean_field_names}()},
\code{\link{get_entity}()},
\code{\link{prep_affiliation}()},
\code{\link{prep_creators}()},
\code{\link{prep_descriptions}()},
\code{\link{prep_for_json}()},
\code{\link{prep_from_metadata_template}()},
\code{\link{prep_fundingReferences}()},
\code{\link{prep_identifier}()},
\code{\link{prep_language}()},
\code{\link{prep_methodology}()},
\code{\link{prep_methods}()},
\code{\link{prep_nameIdentifiers}()},
\code{\link{prep_object}()},
\code{\link{prep_publicationYear}()},
\code{\link{prep_relatedIdentifiers}()},
\code{\link{prep_rights}()},
\code{\link{prep_subjects}()}
}
\concept{JSON Prep}
