% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{prep_from_metadata_template}
\alias{prep_from_metadata_template}
\title{Prepare metadata from the metadata template}
\usage{
prep_from_metadata_template(
  project_metadata,
  prep_methods_list = prep_methods()
)
}
\arguments{
\item{project_metadata}{Data frame. Should correspond to the structure of the project_metadata_template.csv}

\item{prep_methods_list}{list. Named list of methods where each items is a function to applied to corresponding items in x.Default is \code{prep_methods()}}
}
\value{
Named list ready to be converted to json
}
\description{
Prepare metadata from the metadata template
}
\examples{
\dontrun{
# create
wddsWizard::use_template("project_metadata_template.csv",folder = "data", file_name = "my_project_metadata.csv")
project_metadata  <- read.csv("data/my_project_metadata.csv")

prepped_project_metadata  <- wddsWizard::prep_from_metadata_template(project_metadata)

 project_metadat_json <- jsonlite::toJSON(prepped_project_metadata, pretty = TRUE)
}


}
