% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{prep_language}
\alias{prep_language}
\title{Prep language}
\usage{
prep_language(x)
}
\arguments{
\item{x}{named list, vector, or data.frame of with 1:1 name:value pairs}
}
\value{
an unboxed dataframe with 1 row
}
\description{
Prepare the language property - this should describe the language
of the scholarly work.
}
\examples{

a <- data.frame("language" = "en")

prep_language(a)

}
\seealso{
Other JSON Prep: 
\code{\link{clean_field_names}()},
\code{\link{get_entity}()},
\code{\link{prep_affiliation}()},
\code{\link{prep_atomic}()},
\code{\link{prep_creators}()},
\code{\link{prep_descriptions}()},
\code{\link{prep_for_json}()},
\code{\link{prep_from_metadata_template}()},
\code{\link{prep_fundingReferences}()},
\code{\link{prep_identifier}()},
\code{\link{prep_methodology}()},
\code{\link{prep_methods}()},
\code{\link{prep_nameIdentifiers}()},
\code{\link{prep_object}()},
\code{\link{prep_publicationYear}()},
\code{\link{prep_relatedIdentifiers}()},
\code{\link{prep_rights}()},
\code{\link{prep_subjects}()}
}
\concept{JSON Prep}
