% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{prep_methodology}
\alias{prep_methodology}
\title{Prep methodology for conversion to json}
\usage{
prep_methodology(x)
}
\arguments{
\item{x}{List. methodology component of a list}
}
\value{
properly formatted list
}
\description{
Prep methodology for conversion to json
}
\examples{

\dontrun{
prepped_list <- project_metadata_list_entities
 prepped_list$methodology <- prep_methodology(project_metadata_list_entities$methodology)

 OR

 prepped_list <- purrr::modify_at(project_metadata_list_entities,"methodology",prep_methodology)
}

}
\seealso{
Other JSON Prep: 
\code{\link{clean_field_names}()},
\code{\link{get_entity}()},
\code{\link{prep_affiliation}()},
\code{\link{prep_atomic}()},
\code{\link{prep_creators}()},
\code{\link{prep_descriptions}()},
\code{\link{prep_for_json}()},
\code{\link{prep_from_metadata_template}()},
\code{\link{prep_fundingReferences}()},
\code{\link{prep_identifier}()},
\code{\link{prep_language}()},
\code{\link{prep_methods}()},
\code{\link{prep_nameIdentifiers}()},
\code{\link{prep_object}()},
\code{\link{prep_publicationYear}()},
\code{\link{prep_relatedIdentifiers}()},
\code{\link{prep_rights}()},
\code{\link{prep_subjects}()}
}
\concept{JSON Prep}
