% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{prep_methods}
\alias{prep_methods}
\title{Prepare methods}
\usage{
prep_methods()
}
\value{
list of methods
}
\description{
Collection of methods for preparing data conveniently named to make
preparing easier
}
\examples{

prep_methods()

}
\seealso{
Other JSON Prep: 
\code{\link{clean_field_names}()},
\code{\link{get_entity}()},
\code{\link{prep_affiliation}()},
\code{\link{prep_array}()},
\code{\link{prep_array_objects}()},
\code{\link{prep_atomic}()},
\code{\link{prep_creators}()},
\code{\link{prep_data}()},
\code{\link{prep_descriptions}()},
\code{\link{prep_for_json}()},
\code{\link{prep_from_metadata_template}()},
\code{\link{prep_fundingReferences}()},
\code{\link{prep_identifier}()},
\code{\link{prep_language}()},
\code{\link{prep_methodology}()},
\code{\link{prep_nameIdentifiers}()},
\code{\link{prep_object}()},
\code{\link{prep_publicationYear}()},
\code{\link{prep_relatedIdentifiers}()},
\code{\link{prep_rights}()},
\code{\link{prep_subjects}()},
\code{\link{prep_titles}()}
}
\concept{JSON Prep}
