% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{prep_nameIdentifiers}
\alias{prep_nameIdentifiers}
\title{Prepare Name identifiers}
\usage{
prep_nameIdentifiers(x)
}
\arguments{
\item{x}{Data frame from "creators"}
}
\value{
data frame with a nameIdentifiers column as list
}
\description{
These are Persistent identifiers associated with a creator.
}
\details{
Name identifiers in datacite is an array of objects
with properties "nameIdentifier", "nameIdentifierScheme" ,
and "schemeUri". This function takes the \verb{name identifiers} fields and restructures
as a list within the data frame.
}
\examples{
creator_df <- wddsWizard::becker_project_metadata$creators[[1]]
creator_df_nameID_prepped <- prep_nameIdentifiers(creator_df)

}
\seealso{
Other JSON Prep: 
\code{\link{clean_field_names}()},
\code{\link{get_entity}()},
\code{\link{prep_affiliation}()},
\code{\link{prep_array_objects}()},
\code{\link{prep_atomic}()},
\code{\link{prep_creators}()},
\code{\link{prep_data}()},
\code{\link{prep_descriptions}()},
\code{\link{prep_for_json}()},
\code{\link{prep_from_metadata_template}()},
\code{\link{prep_fundingReferences}()},
\code{\link{prep_identifier}()},
\code{\link{prep_language}()},
\code{\link{prep_methodology}()},
\code{\link{prep_methods}()},
\code{\link{prep_object}()},
\code{\link{prep_publicationYear}()},
\code{\link{prep_relatedIdentifiers}()},
\code{\link{prep_rights}()},
\code{\link{prep_subjects}()},
\code{\link{prep_titles}()}
}
\concept{JSON Prep}
