% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{prep_object}
\alias{prep_object}
\title{Prepare an object}
\usage{
prep_object(x, unbox = FALSE)
}
\arguments{
\item{x}{named vector, list, or data frame}

\item{unbox}{logical Should items be unboxed (not arrays)? Default is FALSE
meaning items will remain as arrays when converted to json.}
}
\value{
List of formatted objects
}
\description{
Converts a named vector, list, or data frame to a list, and
optionally unboxes it, so that its recorded as an object.
}
\details{
Note that unboxing will only work on items where you have 1:1 key value pair. So if you have a dataframe with multiple rows or a list  with
multiple values at a given position, it won't work.
}
\examples{

cars_small <- datasets::cars[1:10, ]

# creates an array of objects where each
# row is an object
cars_small |>
  jsonlite::toJSON(pretty = TRUE)

# creates an object with 2 arrays
prep_object(cars_small) |>
  jsonlite::toJSON(pretty = TRUE)

# this makes no difference
x <- list("hello" = 1:10, "world" = "Earth")

prep_object(x) |>
  jsonlite::toJSON(pretty = TRUE)

}
\seealso{
Other JSON Prep: 
\code{\link{clean_field_names}()},
\code{\link{get_entity}()},
\code{\link{prep_affiliation}()},
\code{\link{prep_array}()},
\code{\link{prep_array_objects}()},
\code{\link{prep_atomic}()},
\code{\link{prep_creators}()},
\code{\link{prep_data}()},
\code{\link{prep_descriptions}()},
\code{\link{prep_for_json}()},
\code{\link{prep_from_metadata_template}()},
\code{\link{prep_fundingReferences}()},
\code{\link{prep_identifier}()},
\code{\link{prep_language}()},
\code{\link{prep_methodology}()},
\code{\link{prep_methods}()},
\code{\link{prep_nameIdentifiers}()},
\code{\link{prep_publicationYear}()},
\code{\link{prep_relatedIdentifiers}()},
\code{\link{prep_rights}()},
\code{\link{prep_subjects}()},
\code{\link{prep_titles}()}
}
\concept{JSON Prep}
