% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{prep_publicationYear}
\alias{prep_publicationYear}
\title{Prepare publication year items}
\usage{
prep_publicationYear(x)
}
\arguments{
\item{x}{Named vector, data frame, or list}
}
\value{
an unboxed dataframe with 1 row
}
\description{
wrapper for prep atomic
}
\examples{
pub_year <- data.frame("publicationYear" = "2025")

prep_language(pub_year)


}
\seealso{
Other JSON Prep: 
\code{\link{clean_field_names}()},
\code{\link{get_entity}()},
\code{\link{prep_affiliation}()},
\code{\link{prep_array_objects}()},
\code{\link{prep_atomic}()},
\code{\link{prep_creators}()},
\code{\link{prep_data}()},
\code{\link{prep_descriptions}()},
\code{\link{prep_for_json}()},
\code{\link{prep_from_metadata_template}()},
\code{\link{prep_fundingReferences}()},
\code{\link{prep_identifier}()},
\code{\link{prep_language}()},
\code{\link{prep_methodology}()},
\code{\link{prep_methods}()},
\code{\link{prep_nameIdentifiers}()},
\code{\link{prep_object}()},
\code{\link{prep_relatedIdentifiers}()},
\code{\link{prep_rights}()},
\code{\link{prep_subjects}()},
\code{\link{prep_titles}()}
}
\concept{JSON Prep}
