% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_json.R
\name{prep_relatedIdentifiers}
\alias{prep_relatedIdentifiers}
\title{Prepare related identifiers}
\usage{
prep_relatedIdentifiers(x)
}
\arguments{
\item{x}{data frame with related identifier properties}
}
\value{
List with x marked as do not unbox
}
\description{
Prepare related identifiers
}
\examples{

wddsWizard::becker_project_metadata$relatedIdentifiers |> prep_relatedIdentifiers()

}
