% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_schema.R
\name{process_array_items}
\alias{process_array_items}
\title{Process Array Items}
\usage{
process_array_items(array_items, out)
}
\arguments{
\item{array_items}{list. List of array items for processing.}

\item{out}{data frame. Data frame with name and type.}
}
\value{
data frames with name and type for array items that are objects or character strings atomic (string, null, Boolean, etc) array items.
}
\description{
Processes array items so they can be added to a data frame.
}
\seealso{
Other Schema: 
\code{\link{create_object_list}()},
\code{\link{create_schema_list}()},
\code{\link{datacite_schema}},
\code{\link{disease_data_schema}},
\code{\link{get_ref_list}()},
\code{\link{project_metadata_schema}},
\code{\link{schema_properties}},
\code{\link{schema_terms}},
\code{\link{wdds_schema}}
}
\concept{Schema}
