% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_schema.R
\name{process_array_items}
\alias{process_array_items}
\title{Title}
\usage{
process_array_items(array_items, out)
}
\arguments{
\item{array_items}{list.}

\item{out}{data frame}
}
\value{
data frames with name and type for array items that are objects or character strings atomic (string, null, Boolean, etc) array items.
}
\description{
Title
}
