% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_schema.R
\docType{data}
\name{project_metadata_schema}
\alias{project_metadata_schema}
\title{Wildlife Disease Data Standard}
\format{
An object of class \code{list} of length 6.
}
\usage{
project_metadata_schema
}
\description{
See data standard JSON file for field descriptions.
}
\keyword{datasets}
