% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_schema_versions.R
\name{sanitize_version}
\alias{sanitize_version}
\title{Sanitize version ids}
\usage{
sanitize_version(version)
}
\arguments{
\item{version}{Character. Version identifier.}
}
\value{
Character. Version identifier with no periods.
}
\description{
This function replaces periods with under scores. The different versions
of the data standard are stored in folders with their respective names; however,
having periods in folder names can cause problems on certain operating systems
and makes it more difficult to parse file extensions.
}
\examples{

sanitize_version("v.1.1.0")

}
