% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_required_fields.R
\docType{data}
\name{schema_required_fields}
\alias{schema_required_fields}
\title{Wildlife Disease Data Standard required fields}
\format{
An object of class \code{character} of length 2.
}
\usage{
schema_required_fields
}
\description{
See data standard JSON file for field descriptions. This is a vector of
the required fields for the entire schema.
}
\seealso{
Other Data: 
\code{\link{becker_disease_data}},
\code{\link{becker_project_metadata}},
\code{\link{disease_data_required_fields}},
\code{\link{minimal_disease_data}},
\code{\link{minimal_project_metadata}},
\code{\link{project_metadata_required_fields}},
\code{\link{wdds_to_pharos_map}}
}
\concept{Data}
\keyword{datasets}
