% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_schema.R
\docType{data}
\name{schema_terms}
\alias{schema_terms}
\title{Wildlife Disease Data Standard - schema terms}
\format{
An object of class \code{character} of length 1.
}
\usage{
schema_terms
}
\description{
Markdown of schema terms
}
\seealso{
Other Schema: 
\code{\link{datacite_schema}},
\code{\link{disease_data_schema}},
\code{\link{project_metadata_schema}},
\code{\link{schema_obj}},
\code{\link{schema_properties}},
\code{\link{wdds_schema}}
}
\concept{Schema}
\keyword{datasets}
