% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_schema_versions.R
\name{wdds_json}
\alias{wdds_json}
\title{Provides Access to Versioned Schema Files}
\usage{
wdds_json(version = NULL, file = NULL)
}
\arguments{
\item{version}{Character or NULL. Version of the wdds deposit. Leave as NULL to see
all versions. Default is NULL to return character vector of versions.}

\item{file}{Character or NULL. Specific file from the wdds deposit. Leave as NULL to
see all files in a version. Default is NULL to return character vector of relative file paths.}
}
\value{
Character. Either version identifiers, relative file paths within a version, or a specific file path.
}
\description{
Since schema versions may change during the life cycle of project, it is
important that users have access to all schema versions via this package.
This function allows you to quickly retrieve whichever schema version you
may need.
}
\details{
This function does three things:
\enumerate{
\item Shows all versions of the schema in the package if both version and file are NULL.
\item Provides relative paths to all schema files associated with a version of the schema if only version is provided.
\item Provides a specific file path in a specific version of the schema if version and file path are provided.
}
}
\examples{

# see which versions are in the package

wdds_json()

# see files associated with a version

wdds_json(version = "latest")

# get the file path for a specific file

wdds_json(version = "v_1_0_2",file = "schemas/disease_data.json")


}
\seealso{
Other WDDS deposit: 
\code{\link{batch_download_deposit_versions}()},
\code{\link{download_deposit_version}()},
\code{\link{list_deposit_versions}()},
\code{\link{sanitize_version}()},
\code{\link{set_wdds_version}()},
\code{\link{wdds_data_templates}()},
\code{\link{wdds_example_data}()}
}
\concept{WDDS deposit}
