% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdds_schema.R
\docType{data}
\name{wdds_schema}
\alias{wdds_schema}
\title{Wildlife Disease Data Standard}
\format{
An object of class \code{list} of length 6.
}
\usage{
wdds_schema
}
\description{
See data standard JSON file for field descriptions.
}
\seealso{
Other Schema: 
\code{\link{datacite_schema}},
\code{\link{disease_data_schema}},
\code{\link{project_metadata_schema}},
\code{\link{schema_properties}},
\code{\link{schema_terms}}
}
\concept{Schema}
\keyword{datasets}
